karnak:add some requred changes for clang building its not complete yet but will work on more fixes in the next commit

+++ drivers/misc/mediatek/m4u/mt8163/m4u_mva.c
@@ -10,7 +10,7 @@
 #define MVA_BLOCK_SIZE      (1<<MVA_BLOCK_SIZE_ORDER)	/* 0x40000 */
 #define MVA_BLOCK_ALIGN_MASK (MVA_BLOCK_SIZE-1)	/* 0x3ffff */
 #define MVA_BLOCK_NR_MASK   (MVA_MAX_BLOCK_NR)	/* 0xfff */
-#define MVA_BUSY_MASK       (1<<15)	/* 0x8000 */
+#define MVA_BUSY_MASK       (u8)(1<<15)	/* 0x8000 */
 
 #define MVA_IS_BUSY(index) ((mvaGraph[index]&MVA_BUSY_MASK) != 0)
 #define MVA_SET_BUSY(index) (mvaGraph[index] |= MVA_BUSY_MASK)
